% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_ACO.R
\name{fars_read}
\alias{fars_read}
\title{read a csv text file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character chain containing the name of the file you want to read}
}
\value{
This function returns a dataframe tbl_df
}
\description{
This function reads a csv format file and stores in dataframe tbl_df. For this task uses package "readr" which allows to directly read comressed files
}
\examples{

## fars_read("accident_2013.csv.bz2")

fars_read( system.file("extdata", "accident_2013.csv.bz2", package = "farsUSA") )

}
