% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_ACO.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads several csv text files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list containing the years to estract information}
}
\value{
This function returns a list of tbl_df from all years in "years" with month and year information
}
\description{
This function reads several csv format files and stores month and year into a list of dataframes tbl_df,
containing month and year of the accidents stored, one tbl_df for each csv file.
}
\examples{
fars_read_years(list(2013,2014))

}
