% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_ACO.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes accidents by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years}
}
\value{
This function returns a tbl_df containing number of accidents by month
for the years included in the list parsed as parameter
}
\description{
This function summarizes the number of accidents occurred each month for a list of years
}
\examples{
fars_summarize_years(list(2014,2015))

}
