% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_ACO.R
\name{make_filename}
\alias{make_filename}
\title{Creates a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer indicating the year for filename creation}
}
\value{
This function returns a character variable containing a file name
}
\description{
This function creates the apropriate name of the the file corresponding to accident data for the year 'year'
}
\examples{
make_filename(2013)

}
